requirements.txt
streamlit
sentence-transformers
torch
chromadb
pathlib
semantic_router
groq
python-dotenv
pandas
sqlite3